fw/applib/graphics: fix bitblt overflow for display widths > 255
The `row_bits_left` variable in the tiled bitblt functions was declared
as `uint8_t`, causing silent truncation when `dest_rect.size.w` exceeds
255 pixels. On Gabbro (260x260), this truncates 260 to 4, which triggers
a PBL_ASSERTN crash in `bitblt_bitmap_into_bitmap_tiled_1bit_to_8bit`
when the scroll layer draws its 1-bit shadow bitmaps.

Change the type to `int16_t` to match `dest_rect.size.w` (GRect uses
int16_t for dimensions).

Fixes https://github.com/coredevices/PebbleOS/issues/903

+++ src/fw/applib/graphics/1_bit/bitblt_private.c
@@ -108,7 +108,7 @@ void bitblt_bitmap_into_bitmap_tiled_palette_to_1bit(GBitmap* dest_bitmap,
     RowLookUp look_up = look_ups[dest_y % 2];
 
     int16_t src_x = src_begin_x + src_origin_offset.x;
-    uint8_t row_bits_left = dest_rect.size.w;
+    int16_t row_bits_left = dest_rect.size.w;
     uint32_t *dest_block = (uint32_t *)dest_block_x_begin + (dest_y * dest_row_length_words);
 
     const uint32_t *dest_block_end = dest_block + num_dest_blocks_per_row;
